// apps/web/src/routes/alooid-builder/-data/items-enum.ts
export enum ItemId {
  AMETISTA = "719",
  AQUAMARINA = "720",
  TOPAZ = "728",
  ROSA_ETERNA = "748",
  BLOCO_DE_SUCATA = "6961",
  TANQUE_VELHO = "6962",
  ARMA_DANIFICADA = "25668",
  SOMBRIODECON = "25729",
  ZELUNIUM = "25731",
  ARQUIVOS_DE_BIOPESQUISA = "25786",
  FRAGMENTO_DE_EXPERIMENTO = "25787",
  DYNITE = "25814",
  TICKET_SOMBRIO_GERAL = "30060",
  TICKET_SOMBRIO_HABILIDADES = "30061",
  TICKET_SOMBRIO_CLASSES = "30062",
  TICKET_SOMBRIO_AVANCADO = "30064",
  CAIXA_AREIA_ESTELAR_DE_BRUXA_400 = "40092",
  MOEDA_DO_INICIANTE = "50169",
  CAIXA_ANCESTRAL = "51115",
  AUTOMATIC_MODULE_BOX = "100160",
  BARMEAL_TICKET = "1000103",
  PIECE_OF_SIN = "1000257",
  FRAGMENT_OF_GOOD_WILL = "1000263",
  AMBAR = "1000321",
  PO_DE_ETER = "1000322",
  GEMA_ETER = "1000323",
  MINNEAS = "1000367",
}